backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pyyaml>=6.0.1
elasticsearch>=8.0.0
mysql-connector-python>=8.0.0
pymongo>=4.0.0
dnspython>=2.4.0
python-multipart>=0.0.6
aiofiles>=23.0.0
